added authorisation to courses

+++ controllers/bootcamps.js
@@ -29,6 +29,17 @@ exports.getBootcamp = asyncHandler(async (req, res, next) => {
 // @route: POST /api/v1/bootcamps
 // @access: Private
 exports.createBootcamp = asyncHandler(async (req, res, next) => {
+    req.body.user = req.user.id;
+
+
+    const publishedBootcamp = await Bootcamp.findOne({user: req.user.id});
+
+
+    // If the user is not an admin, they can only add one bootcamp
+    if(publishedBootcamp && req.user.role !== "admin") {
+        return next(new ErrorResponse(`The user with ID ${req.user.id} has already published a course`, 400))
+    }
+
     const bootcamp = await Bootcamp.create(req.body);
     res.status(201).json({success: true, data: bootcamp})
 });
@@ -37,17 +48,26 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 // @route: PUT /api/v1/bootcamps/:id
 // @access: Private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-    const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-        new: true,
-        runValidators: true
-    });
+
+    const bootcamp = await Bootcamp.findById(req.params.id);
 
 
     if(!bootcamp) {
         return next(new ErrorResponse(`Resource could not be updated.`, 400));
     }
 
-    res.status(200).json({success: true, data: bootcamp})
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(
+            new ErrorResponse(`User ${req.params.id} is not authorised to update this bootcamp`, 401)
+        )
+    }
+
+    const bootcampUpdate = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
+        new: true,
+        runValidators: true
+    });
+
+    res.status(200).json({success: true, data: bootcampUpdate})
 });
 
 // @desc: Delete bootcamp
@@ -59,6 +79,12 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
 
     if(!bootcamp) {
         return next(new ErrorResponse(`DB Error.`, 400));
+    }
+
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(
+            new ErrorResponse(`User ${req.params.id} is not authorised to delete this bootcamp`, 401)
+        )
     }
 
     await bootcamp.deleteOne({id: req.params.id});
@@ -95,6 +121,12 @@ exports.getBootcampInRadius = asyncHandler(async (req, res, next) => {
 exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
     const bootcamp = await Bootcamp.findById(req.params.id);
 
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+        return next(
+            new ErrorResponse(`User ${req.params.id} is not authorised to update this bootcamp`, 401)
+        )
+    }
+
 
     if(!bootcamp) {
         return next(new ErrorResponse(`Resource could not be updated.`, 400));
@@ -116,7 +148,6 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
 
     // Create custom filename
     file.name = `photo_${bootcamp._id}${path.parse(file.originalname).ext}`;
-console.log("file: ", file)
     await Bootcamp.findByIdAndUpdate(req.params.id, {photo: file.name});
 
     res.status(200).json({
